+++ packages/deploy/test/stateTransform.test.ts
@@ -90,6 +90,46 @@ test('toNextState adding a job', (t) => {
     name: 'project-name',
     description: 'my test project',
     project_credentials: {},
+    collections: {},
+  });
+});
+
+test('toNextState deleting a credential', (t) => {
+  const spec = {
+    name: 'project-name',
+    description: 'my test project',
+    collections: {},
+    workflows: {},
+  };
+
+  const state = {
+    workflows: {},
+    project_credentials: {},
+    collections: {
+      'test-collection': {
+        id: 'f8e1c1e1-5c5a-4d9b-8e9f-4b2f6b1c2f4e',
+        name: 'test-collection',
+      },
+    },
+    id: 'ecb683d1-5e5a-4c4f-9165-e143e2eeeb48',
+    name: 'project-name',
+    description: 'my test project',
+  };
+
+  let result = mergeSpecIntoState(state, spec);
+
+  t.deepEqual(result, {
+    workflows: {},
+    id: 'ecb683d1-5e5a-4c4f-9165-e143e2eeeb48',
+    name: 'project-name',
+    description: 'my test project',
+    project_credentials: {},
+    collections: {
+      'test-collection': {
+        id: 'f8e1c1e1-5c5a-4d9b-8e9f-4b2f6b1c2f4e',
+        delete: true,
+      },
+    },
   });
 });
 
@@ -117,6 +157,7 @@ test('toNextState with empty state', (t) => {
     name: 'my project',
     description: 'some helpful description',
     project_credentials: {},
+    collections: {},
     workflows: {
       'workflow-one': {
         id: jp.query(result, '$..workflows["workflow-one"].id')[0],
@@ -176,6 +217,7 @@ test('toNextState with no changes', (t) => {
     name: 'my project',
     description: 'for the humans',
     project_credentials: {},
+    collections: {},
     workflows: {
       'workflow-one': {
         id: '8124e88c-566f-472f-be38-363e588af55a',
@@ -291,6 +333,7 @@ test('toNextState with a new job', (t) => {
     name: 'my project',
     description: 'some other description',
     project_credentials: {},
+    collections: {},
     workflows: {
       'workflow-one': {
         id: '8124e88c-566f-472f-be38-363e588af55a',
@@ -538,6 +581,7 @@ test('getStateFromProjectPayload with minimal project', (t) => {
     id: 'xyz',
     name: 'project',
     project_credentials: {},
+    collections: {},
     workflows: {
       a: {
         id: 'wf-a',